tests/wpt/tests/css/css-sizing/stretch/indefinite-2.html

<!DOCTYPE html>
<link rel="help"
  href="https://drafts.csswg.org/css-sizing-4/#stretch-fit-sizing">
<link rel="help" href="https://github.com/w3c/csswg-drafts/issues/11006">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<meta name="assert"
  content="with indefinite available space, min-height:stretch resolves to 0">

<body
  onload="checkLayout('[data-expected-client-height]')">

  <div style="min-height: stretch; height: 0; width: 200px; border: solid"
    data-expected-client-height="0">lorem ipsum</div>
